host-bridge/requirements.txt
flask
mss
pygetwindow
pyautogui
pillow
